+++ spec/spec_helper.rb
@@ -6,7 +6,8 @@ require 'generator_spec'
 
 Lite::Ruby.configure do |config|
   config.monkey_patches = %w[
-    array date enumerable hash integer kernel numeric object open_struct range string struct time
+    array boolean date enumerable hash integer kernel numeric object open_struct range string
+    struct time
   ]
 end