+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.4] - 2019-08-09
+## [1.0.4] - 2019-08-010
+### Changed
+- Improved speed for Array denillify
+- Improved speed for Hash denillify
+
+## [1.0.4] - 2019-08-010
 ### Added
 - Added more boolean value checks
 - Added Numeric => `markdown_percentage`
@@ -15,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Object => `to_bool`
 - Added String => `dedupe`
 ### Changed
-- Move boolean classes to its own file
+- Move Boolean classes to its own file
 
 ## [1.0.3] - 2019-07-30
 ### Added